+++ src/main/java/com/codeit/mission/discodeit/service/basic/BasicUserService.java
@@ -8,6 +8,7 @@ import com.codeit.mission.discodeit.entity.BinaryContent;
 import com.codeit.mission.discodeit.entity.User;
 import com.codeit.mission.discodeit.entity.UserStatus;
 import com.codeit.mission.discodeit.mapper.UserMapper;
+import com.codeit.mission.discodeit.repository.BinaryContentRepository;
 import com.codeit.mission.discodeit.repository.UserRepository;
 import com.codeit.mission.discodeit.repository.UserStatusRepository;
 import com.codeit.mission.discodeit.service.UserService;
@@ -23,17 +24,18 @@ import org.springframework.transaction.annotation.Transactional;
 
 @RequiredArgsConstructor
 @Service
-@Transactional
 public class BasicUserService implements UserService {
 
     private final UserRepository userRepository;
     private final UserStatusRepository userStatusRepository;
-    private final BinaryContentStorage binaryContentStorage;
     private final UserMapper userMapper;
+    private final BinaryContentRepository binaryContentRepository;
+    private final BinaryContentStorage binaryContentStorage;
 
+    @Transactional
     @Override
-    public User create(UserCreateRequest userCreateRequest,
-        Optional<BinaryContentCreateRequest> optionalProfileCreateRequest) {
+    public UserDto create(UserCreateRequest userCreateRequest,
+            Optional<BinaryContentCreateRequest> optionalProfileCreateRequest) {
         String username = userCreateRequest.username();
         String email = userCreateRequest.email();
 
@@ -42,52 +44,54 @@ public class BasicUserService implements UserService {
         }
         if (userRepository.existsByUsername(username)) {
             throw new IllegalArgumentException(
-                "User with username " + username + " already exists");
+                    "User with username " + username + " already exists");
         }
 
-        BinaryContent profile = optionalProfileCreateRequest
-            .map(profileRequest -> {
-                String fileName = profileRequest.fileName();
-                String contentType = profileRequest.contentType();
-                byte[] bytes = profileRequest.bytes();
-
-                BinaryContent binaryContent = new BinaryContent(fileName, (long) bytes.length,
-                    contentType);
-                binaryContentStorage.put(binaryContent.getId(), bytes);
-
-                return binaryContent;
-            })
-            .orElse(null);
-
+        BinaryContent nullableProfile = optionalProfileCreateRequest
+                .map(profileRequest -> {
+                    String fileName = profileRequest.fileName();
+                    String contentType = profileRequest.contentType();
+                    byte[] bytes = profileRequest.bytes();
+                    BinaryContent binaryContent = new BinaryContent(fileName, (long) bytes.length,
+                            contentType);
+                    binaryContentRepository.save(binaryContent);
+                    binaryContentStorage.put(binaryContent.getId(), bytes);
+                    return binaryContent;
+                })
+                .orElse(null);
         String password = userCreateRequest.password();
-        User user = new User(username, email, password, profile);
-        User createdUser = userRepository.save(user);
+
+        User user = new User(username, email, password, nullableProfile);
         Instant now = Instant.now();
-        UserStatus userStatus = new UserStatus(createdUser, now);
-        userStatusRepository.save(userStatus);
-        return createdUser;
+        UserStatus userStatus = new UserStatus(user, now);
+
+        userRepository.save(user);
+        return userMapper.toDto(user);
     }
 
     @Override
     public UserDto find(UUID userId) {
         return userRepository.findById(userId)
-            .map(userMapper::toDto)
-            .orElseThrow(() -> new NoSuchElementException("User with id " + userId + " not found"));
+                .map(userMapper::toDto)
+                .orElseThrow(
+                        () -> new NoSuchElementException("User with id " + userId + " not found"));
     }
 
     @Override
     public List<UserDto> findAll() {
-        return userRepository.findAll()
-            .stream()
-            .map(userMapper::toDto)
-            .toList();
+        return userRepository.findAllWithProfileAndStatus()
+                .stream()
+                .map(userMapper::toDto)
+                .toList();
     }
 
+    @Transactional
     @Override
-    public User update(UUID userId, UserUpdateRequest userUpdateRequest,
-        Optional<BinaryContentCreateRequest> optionalProfileCreateRequest) {
+    public UserDto update(UUID userId, UserUpdateRequest userUpdateRequest,
+            Optional<BinaryContentCreateRequest> optionalProfileCreateRequest) {
         User user = userRepository.findById(userId)
-            .orElseThrow(() -> new NoSuchElementException("User with id " + userId + " not found"));
+                .orElseThrow(
+                        () -> new NoSuchElementException("User with id " + userId + " not found"));
 
         String newUsername = userUpdateRequest.newUsername();
         String newEmail = userUpdateRequest.newEmail();
@@ -96,33 +100,35 @@ public class BasicUserService implements UserService {
         }
         if (userRepository.existsByUsername(newUsername)) {
             throw new IllegalArgumentException(
-                "User with username " + newUsername + " already exists");
+                    "User with username " + newUsername + " already exists");
         }
 
-        BinaryContent newProfile = optionalProfileCreateRequest
-            .map(profileRequest -> {
-                String fileName = profileRequest.fileName();
-                String contentType = profileRequest.contentType();
-                byte[] bytes = profileRequest.bytes();
+        BinaryContent nullableProfile = optionalProfileCreateRequest
+                .map(profileRequest -> {
 
-                BinaryContent binaryContent = new BinaryContent(fileName, (long) bytes.length,
-                    contentType);
-                binaryContentStorage.put(binaryContent.getId(), bytes);
-
-                return binaryContent;
-            })
-            .orElse(user.getProfile());
+                    String fileName = profileRequest.fileName();
+                    String contentType = profileRequest.contentType();
+                    byte[] bytes = profileRequest.bytes();
+                    BinaryContent binaryContent = new BinaryContent(fileName, (long) bytes.length,
+                            contentType);
+                    binaryContentRepository.save(binaryContent);
+                    binaryContentStorage.put(binaryContent.getId(), bytes);
+                    return binaryContent;
+                })
+                .orElse(null);
 
         String newPassword = userUpdateRequest.newPassword();
-        user.update(newUsername, newEmail, newPassword, newProfile);
+        user.update(newUsername, newEmail, newPassword, nullableProfile);
 
-        return user;
+        return userMapper.toDto(user);
     }
 
+    @Transactional
     @Override
     public void delete(UUID userId) {
-        User user = userRepository.findById(userId)
-            .orElseThrow(() -> new NoSuchElementException("User with id " + userId + " not found"));
+        if (userRepository.existsById(userId)) {
+            throw new NoSuchElementException("User with id " + userId + " not found");
+        }
 
         userRepository.deleteById(userId);
     }
